Fix error of __list_find The node on the linked list should follow the ordering by its key. However, the function list_insert doesn't actually do this because of the wrong implementation of __list_find.
There are three main fixes in this patch. First,
once we should do synchronization, we have to iterate the list
again but not return from the function. Second, the while loop should
iterate until the last node but not the second last one. Finally,
under the normal situation that the tail node with key UINTPTR_MAX should
not be deleted (unless the end of the program), some codes are never executed
which can be removed.